RAG_chat_openAI/requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
PyPDF2
openai